+++ src/linux_arctis_manager/gui/main_app.py
@@ -10,6 +10,7 @@ from linux_arctis_manager.gui.base_app import QBaseDesktopApp
 from linux_arctis_manager.gui.dbus_wrapper import DbusWrapper
 from linux_arctis_manager.gui.main_app_proto_widget import QMainAppProtoWidget
 from linux_arctis_manager.gui.settings_widget import QSettingsWidget
+from linux_arctis_manager.gui.systray_toggles_widget import QSystrayTogglesWidget
 from linux_arctis_manager.gui.status_widget import QStatusWidget
 from linux_arctis_manager.gui.ui_utils import get_icon_pixmap
 from linux_arctis_manager.i18n import I18n
@@ -44,11 +45,13 @@ class QMainApp(QBaseDesktopApp):
         self.status_widget = QStatusWidget(self.main_panel)
         self.general_settings_widget = QSettingsWidget(self.main_panel, 'general', 'general')
         self.device_settings_widget = QSettingsWidget(self.main_panel, 'device', 'device')
+        self.systray_toggles_widget = QSystrayTogglesWidget(self.main_panel)
 
         self.main_panel_widgets: dict[str, QWidget] = {
             'status': self.status_widget,
             'general': self.general_settings_widget,
             'device': self.device_settings_widget,
+            'systray_toggles': self.systray_toggles_widget,
         }
 
         for widget in self.main_panel_widgets.values():
@@ -58,6 +61,7 @@ class QMainApp(QBaseDesktopApp):
         self.dbus_wrapper.sig_status.connect(self.status_widget.update_status)
         self.dbus_wrapper.sig_settings.connect(self.general_settings_widget.update_settings)
         self.dbus_wrapper.sig_settings.connect(self.device_settings_widget.update_settings)
+        self.dbus_wrapper.sig_settings.connect(self.systray_toggles_widget.update_settings)
 
         self.switch_panel('status')
         self.dbus_wrapper.start()
@@ -99,6 +103,7 @@ class QMainApp(QBaseDesktopApp):
             ('status', I18n.get_instance().translate('ui', 'status')),
             ('general', I18n.get_instance().translate('ui', 'general')),
             ('device', I18n.get_instance().translate('ui', 'device')),
+            ('systray_toggles', I18n.get_instance().translate('ui', 'systray_toggles')),
         ]
 
         for value, text in self.side_panel_items:
@@ -118,7 +123,7 @@ class QMainApp(QBaseDesktopApp):
 
         return window
     
-    def switch_panel(self, panel: Literal['status', 'general', 'device']) -> None:
+    def switch_panel(self, panel: Literal['status', 'general', 'device', 'systray_toggles']) -> None:
         if not self.main_panel_widgets[panel].isHidden():
             return